--- src/components/CreateTaskButton/index.ts
@@ -1,1 +1,0 @@
-export { CreateTaskButton } from "./CreateTaskButton"
