+++ src/components/FormNewTask/index.ts
@@ -1,0 +1,1 @@
+export { FormNewTask as NewTaskForm } from "./FormNewTask"
