--- src/components/NewTaskForm/index.ts
@@ -1,1 +1,0 @@
-export { NewTaskForm } from "./NewTaskForm"
